google auth working now
